+++ apps/docs/src/content/04-components/upload/image-cropper/examples/aspect.tsx
@@ -1,0 +1,6 @@
+import { ImageCropper } from "@mittwald/flow-react-components";
+
+<ImageCropper
+  image="https://mittwald.github.io/flow/assets/mittwald_logo_rgb.jpg"
+  aspectRatio={16 / 9}
+/>;