+++ packages/components/src/components/Slider/stories/Default.stories.tsx
@@ -26,7 +26,7 @@ export const ReadOnly: Story = {
   args: { isReadOnly: true },
 };
 
-export const WithUnit: Story = {
+export const WithFormatOptions: Story = {
   render: (props) => (
     <Slider
       {...props}